% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nepse_data.R
\docType{data}
\name{nepse_index_id}
\alias{nepse_index_id}
\title{Stock Index ID for Index and Sub-index of NEPSE.}
\format{A data frame with two variables:
\describe{
\item{index_name}{Name of index or sub-index}
\item{index_id}{ID of index or sub_index which ranges from 51 to 63}
}}
\source{
\url{http://www.nepalstock.com.np/}
}
\usage{
nepse_index_id
}
\description{
The data consists of index name and index id of Nepal Stock Exchange.
These are from \url{http://www.nepalstock.com.np/}.
}
\keyword{datasets}

