% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nepse_data.R
\docType{data}
\name{nepse_stock_symbols}
\alias{nepse_stock_symbols}
\title{Stock Details for Listed Companies in NEPSE.}
\format{A data frame with five variables:
\describe{
\item{company_name}{Name of the company listed in the NEPSE}
\item{stock_type}{Stock type of a listed company: ordinary share or
non-ordinary share}
\item{stock_symbol}{Stock symbol of the company listed in the NEPSE}
\item{stock_id}{Stock id of the company listed in the NEPSE}
\item{isin}{international securities identification number}
}}
\usage{
nepse_stock_symbols
}
\description{
The data consists of stock symbols, stock types, stock ids, and international
 securities identification number (isin) for all listed companies in Nepal
 Stock Exchange. Stock symbols, stock types, and stock ids are from
 \url{http://www.nepalstock.com.np/} .isin is from
 \url{http://www.cdscnp.com/}. Note that there are some missings for isin, as
some companies still have not obtained it.
}
\keyword{datasets}

